+++ services/core-web/src/constants/styles.js
@@ -9,28 +9,3 @@ export const COLOR = {
   dangerButtonRed: "#BC2929",
   backgroundWhite: "#ffffff",
 };
-
-// Styling types for various document/application status codes
-
-const undefinedStatusStyleType = "default";
-
-const noticeOfWorkStatusStyleTypes = {
-  Accepted: "success",
-  "Under Review": "processing",
-  Withdrawn: "warning",
-};
-
-const varianceApplicationStatusStyleType = {
-  Approved: "success",
-  Denied: "error",
-  "Not Applicable": "default",
-  "In Review": "processing",
-  "Ready for Decision": "processing",
-  Withdrawn: "warning",
-};
-
-export const getNoticeOfWorkApplicationStatusStyleType = (status) =>
-  noticeOfWorkStatusStyleTypes[status] || undefinedStatusStyleType;
-
-export const getVarianceApplicationStatusStyleType = (status) =>
-  varianceApplicationStatusStyleType[status] || undefinedStatusStyleType;
